Fixed random GPU crash in Vulkan which could also have effected d3d12 renderer, tweaked shader tool to take the shader formats it want’s as input and switched default shaders to all be crosscompiled from the same hlsl source shaders.

+++ Source/Rendering/RNShader.cpp
@@ -198,7 +198,7 @@ namespace RN
 		else if (name->IsEqual(RNCSTR("lights_directionalcount")) || name->IsEqual(RNCSTR("directionalLightsCount")))
 		{
 			_identifier = DirectionalLightsCount;
-			_type = PrimitiveType::Int16;
+			_type = PrimitiveType::Uint32;
 		}
 		else if (name->IsEqual(RNCSTR("lights_directional")) || name->IsEqual(RNCSTR("directionalLights")))
 		{
@@ -213,7 +213,7 @@ namespace RN
 		else if (name->IsEqual(RNCSTR("shadows_directional_matrices_count")) || name->IsEqual(RNCSTR("directionalShadowMatricesCount")))
 		{
 			_identifier = DirectionalShadowMatricesCount;
-			_type = PrimitiveType::Int16;
+			_type = PrimitiveType::Uint32;
 		}
 		else if (name->IsEqual(RNCSTR("shadows_directional_info")) || name->IsEqual(RNCSTR("directionalShadowInfo")))
 		{
